src/main/webapp/app/entities/fluxo-execucao/delete/fluxo-execucao-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { FluxoExecucaoService } from '../service/fluxo-execucao.service';

import { FluxoExecucaoDeleteDialogComponent } from './fluxo-execucao-delete-dialog.component';

describe('FluxoExecucao Management Delete Component', () => {
  let comp: FluxoExecucaoDeleteDialogComponent;
  let fixture: ComponentFixture<FluxoExecucaoDeleteDialogComponent>;
  let service: FluxoExecucaoService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [FluxoExecucaoDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(FluxoExecucaoDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(FluxoExecucaoDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(FluxoExecucaoService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
